requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
python-dotenv==1.0.0
Flask-SQLAlchemy==3.1.1
gunicorn==21.2.0
